PET-111 refactor : 테스트 패키지 이름 수정
